REFACTOR: NAV-16 - Refactoring suggestions
- Avoid boiler plate code, use lombok instead.
- Add REST test file.
- Rename api to app package, due to architectural overview.
