--- src/main/java/ch/naviqore/api/model/Connection.java
@@ -1,58 +1,0 @@
-package ch.naviqore.api.model;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.Getter;
-
-import java.util.List;
-import java.util.Objects;
-
-/**
- * Connection
- */
-@Getter
-public class Connection {
-
-    @JsonProperty("legs")
-    private List<Leg> legs;
-
-    public Connection(List<Leg> legs) {
-        this.legs = legs;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        Connection connection = (Connection) o;
-        return Objects.equals(this.legs, connection.legs);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(legs);
-    }
-
-    @Override
-    public String toString() {
-        StringBuilder sb = new StringBuilder();
-        sb.append("class Connection {\n");
-        sb.append("    legs: ").append(toIndentedString(legs)).append("\n");
-        sb.append("}");
-        return sb.toString();
-    }
-
-    /**
-     * Convert the given object to string with each line indented by 4 spaces (except the first line).
-     */
-    private String toIndentedString(Object o) {
-        if (o == null) {
-            return "null";
-        }
-        return o.toString().replace("\n", "\n    ");
-    }
-}
-
