--- src/main/java/ch/naviqore/api/model/Coordinate.java
@@ -1,63 +1,0 @@
-package ch.naviqore.api.model;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.Getter;
-
-import java.util.Objects;
-
-/**
- * Coordinate
- */
-@Getter
-public class Coordinate {
-
-    @JsonProperty("latitude")
-    private double latitude;
-
-    @JsonProperty("longitude")
-    private double longitude;
-
-    public Coordinate(double latitude, double longitude) {
-        this.latitude = latitude;
-        this.longitude = longitude;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        Coordinate coordinate = (Coordinate) o;
-        return Objects.equals(this.latitude, coordinate.latitude) && Objects.equals(this.longitude,
-                coordinate.longitude);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(latitude, longitude);
-    }
-
-    @Override
-    public String toString() {
-        StringBuilder sb = new StringBuilder();
-        sb.append("class Coordinate {\n");
-        sb.append("    latitude: ").append(toIndentedString(latitude)).append("\n");
-        sb.append("    longitude: ").append(toIndentedString(longitude)).append("\n");
-        sb.append("}");
-        return sb.toString();
-    }
-
-    /**
-     * Convert the given object to string with each line indented by 4 spaces (except the first line).
-     */
-    private String toIndentedString(Object o) {
-        if (o == null) {
-            return "null";
-        }
-        return o.toString().replace("\n", "\n    ");
-    }
-}
-
