+++ src/main/java/ch/naviqore/app/model/DistanceToStop.java
@@ -1,0 +1,18 @@
+package ch.naviqore.app.model;
+
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import lombok.ToString;
+
+@RequiredArgsConstructor
+@EqualsAndHashCode
+@ToString
+@Getter
+public class DistanceToStop {
+
+    private final Stop stop;
+    private final double distance;
+
+}
+
