+++ src/main/java/ch/naviqore/app/model/Leg.java
@@ -1,0 +1,29 @@
+package ch.naviqore.app.model;
+
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import lombok.ToString;
+import org.springframework.format.annotation.DateTimeFormat;
+
+import java.time.LocalDateTime;
+
+@RequiredArgsConstructor
+@EqualsAndHashCode
+@ToString
+@Getter
+public class Leg {
+
+    private final Coordinate from;
+    private final Coordinate to;
+    private final Stop fromStop;
+    private final Stop toStop;
+    private final LegType type;
+    @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME)
+    private final LocalDateTime departureTime;
+    @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME)
+    private final LocalDateTime arrivalTime;
+    private final Trip trip;
+
+}
+
