+++ src/main/java/ch/naviqore/app/model/LegType.java
@@ -1,0 +1,33 @@
+package ch.naviqore.app.model;
+
+import com.fasterxml.jackson.annotation.JsonCreator;
+import com.fasterxml.jackson.annotation.JsonValue;
+import lombok.AccessLevel;
+import lombok.RequiredArgsConstructor;
+import lombok.ToString;
+
+@RequiredArgsConstructor(access = AccessLevel.PRIVATE)
+@ToString
+public enum LegType {
+
+    WALK("WALK"),
+    ROUTE("ROUTE");
+
+    private final String value;
+
+    @JsonCreator
+    public static LegType fromValue(String value) {
+        for (LegType b : LegType.values()) {
+            if (b.value.equals(value)) {
+                return b;
+            }
+        }
+        throw new IllegalArgumentException("Unexpected value '" + value + "'");
+    }
+
+    @JsonValue
+    public String getValue() {
+        return value;
+    }
+}
+
