+++ src/main/java/ch/naviqore/app/model/StopTime.java
@@ -1,0 +1,24 @@
+package ch.naviqore.app.model;
+
+import lombok.AllArgsConstructor;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.ToString;
+import org.springframework.format.annotation.DateTimeFormat;
+
+import java.time.LocalDateTime;
+
+@AllArgsConstructor
+@EqualsAndHashCode
+@ToString
+@Getter
+public class StopTime {
+
+    private Stop stop;
+    @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME)
+    private LocalDateTime arrivalTime;
+    @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME)
+    private LocalDateTime departureTime;
+
+}
+
